+++ src/entities/post/createPost/createPost.api.ts
@@ -1,5 +1,5 @@
-import { http } from "../../../shared/api"
-import { CreatePostResponse, NewPost } from "./createPost.model"
+import { http } from "@shared/api"
+import { CreatePostResponse, NewPost } from "@entities/post/createPost/createPost.model"
 
 export const createPost = async (post: NewPost): Promise<CreatePostResponse> => {
   return await http.post("/api/posts/add", {
