+++ src/entities/post/deletePost/index.ts
@@ -1,1 +1,2 @@
 export * from "./deletePost.api"
+export * from "./deletePost.model"
