+++ src/entities/post/getPosts/index.ts
