+++ src/entities/post/index.ts
@@ -1,0 +1,5 @@
+export * from "./createPost"
+export * from "./deletePost"
+export * from "./getPosts"
+export * from "./model"
+export * from "./updatePost"
