+++ src/features/post/getPosts/index.ts
@@ -1,0 +1,1 @@
+export * from "./ui"
