+++ src/features/post/getPosts/ui/PostDetail.tsx
@@ -1,6 +1,7 @@
-import { Post } from "../../../../entities/post/model"
-import { Dialog, DialogContent, DialogHeader, DialogTitle, HighlightText } from "../../../../shared/ui"
-import { CommentList } from "../../../comment/getComment/ui/CommentList"
+import { Post } from "@entities/post"
+import { Dialog, DialogContent, DialogHeader, DialogTitle, HighlightText } from "@shared/ui"
+import { CommentList } from "@features/comment/getComment/ui/CommentList"
+
 interface PostDetailProps {
   isOpen: boolean
   close: () => void
