+++ src/features/search/ui/Pagination.tsx
@@ -1,4 +1,4 @@
-import { Button, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../../shared/ui"
+import { Button, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui"
 import { usePagination } from "../model/usePagination"
 
 interface PaginationProps {
