+++ src/features/search/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./Pagination"
+export * from "./SearchBar"
