+++ src/widgets/post/ui/PostDashboard.tsx
@@ -1,6 +1,6 @@
 import { FormProvider } from "react-hook-form"
 import { CreatePostFormButton } from "../../../features/post/createPost/ui/CreatePostFormButton"
-import { PostTable } from "../../../features/post/getPost/ui/PostTable"
+import { PostTable } from "../../../features/post/getPosts/ui/PostTable"
 import { useSearchBarForm } from "../../../features/search/model/useSearchBarForm"
 import { SearchBar } from "../../../features/search/ui/SearchBar"
 import { Card, CardContent, CardHeader, CardTitle } from "../../../shared/ui"
